requirements.txt
aiogram==3.7.0
SQLAlchemy==2.0.29
asyncpg==0.29.0
psycopg2-binary==2.9.9

httpx==0.27.0
APScheduler==3.10.4

python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1

alembic==1.13.1
